Use tagNameFormat to include correct tag in released pom file (#943)
The 933.0.1 pom file stored in the Jenkins artifact repository includes
an invalid scm.tag value because the tagNameFormat property was removed
on the 933.x branch.

https://maven.apache.org/maven-release/maven-release-plugin/prepare-mojo.html#tagNameFormat
Docs say the default tagNameFormat is '@{project.artifactId}-@{project.version}'.
That matches with the value that is in the pom file,
'cloudbees-bitbucket-branch-source-933.0.1'.

Download the release 933.0.1 pom file from
https://repo.jenkins-ci.org/artifactory/releases/org/jenkins-ci/plugins/cloudbees-bitbucket-branch-source/933.0.1/cloudbees-bitbucket-branch-source-933.0.1.pom

The 933.0.1 release cannot be included in the Jenkins plugin BOM because
of that invalid scm.tag.  A 933.0.2 release is needed after this change
is merged so that the released pom file includes the tag value that
matches the tag written to the repository by the Apache Maven release plugin.