[JENKINS-64418] Add exponential backoff to BitBucket rate limit retry loop (#959)
Fix duplicate base URL in request to Bitbucket Server. The problem is that getSingleBranch prepends this.baseURL to the URI
template and calls getResource, which expands the template and prepends this.baseURL to the resulting URL again.
Method getResource is called only by getSingleTag and getSingleBranch, and both of them include the base URL in the template. Thus, getResource should not prepend it again.