Add a new trait to discard all branches with head commits older than the configured days.
This trait is useful when you do not want to pollute the Jenkins instance with older work such as support, release or feature branches for which there will be no further work, for which without this trait the work will be kept there for weeks or years taking up resources. When a new commit is pushed to the discarded branch, a new Jenkins work will be recreated.