[JENKINS-75729] hookSignatureCredentialsId in a JCasC configuration sometimes is removed (#1053)
Add a new constructor to save hook signature settings. Deserialization uses reflection and the order in which setters are called is not guaranteed, this means that for example in JCasC the hookSignatureCredentialsId may not be set if enableHookSignature is not called as first.
Add test case for hook signature in a JCasC configuration