[JENKINS-75676] Client certificate credentials not handled properly (#1072)
Add a custom TlsSocketStrategy implementation to retrieve an SSLContext from a specific HttpContext or delegate it to DefaultClientTlsSocketStrategy if it is not present. This way, the SSLContext created from a keyStore (client certificate) is configured to be used by the underlying HttpConnection of the shared connection pool. The HttpContext is created each time a new Bitbucket client is instantiated, which means that the HttpContext is bound to the lifecycle of the client instance.